meta-ktn: ktn.conf: Include poky.conf and make Yocto-base-series-agnostic
In the past we copied content from poky.conf into our own distro
config. This bears the problem of redundant configuration. Let's
include poky.conf and only overwrite values where needed.

If we need a custom distro config based on ktn.conf (for customer
projects, etc.) we should create a copy of ktn.conf in the custom
layer and change the content as needed.

This also makes use of the BASE_SERIES variable passed via
local.conf to determine for which Yocto series/release we are
building.