autostart-app: Use drm interface for STM32MP1
Use drm interface for Qt backend linuxfb. This fixes some bugs when scaling
some UI elements. These bugs are visible with the 'gallery' app on pages
with choices: The circles are misaligned and misscaled when 'QT_QPA_FB_DRM'
is not set. If it is set, all elements are error free.
Currently this setting is only activated for the STM32MP1 platform. Maybe this would be sensible for the imx platforms also. Enabling this for the imx
platforms needs currently some testing.