kas: Restructure config files and use env variable to set Yocto base release/series
In order to support multiple Yocto releases and switch between them,
use an env variable YOCTO_SERIES to set the default refspec. This
value is also passed to Bitbake in the BASE_SERIES variable.

With this we can create series conf files like dunfell.yml to select
a certain Yocto base release/series.