Replace occurences of 'ktn'/'KTN' with 'ked'/'KED' in layer names and other places
In the future we will reference the BSP as meta-ked-bsp. Adjust names
of layers, files, etc. accordingly. Only some OE metadata like image,
distro or machine override names are keeping the 'ktn' for now.

KED stands for Kontron Electronics Deutschland GmbH where the BSP
is maintained. Nowadays it support only products of the KED branch
and products of the Kontron Europe GmbH or other branches are not
supported.